pyproject.toml
[tool.poetry]
name = "echonextdi"
version = "0.1.0"
description = "Light and intuitive Dependency Injection Library for pyEchoNext Project"
authors = ["alexeev-prog <alexeev.dev@mail.ru>"]
license = "GNU LGPL 2.1"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
orjson = "^3.10.12"
pyyaml = "^6.0.2"
toml = "^0.10.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"
pytest = "^8.3.4"
ruff = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
